Allow Sensor Data records to use 6-bit ASCII encoding
IPMI sensor names are permitted to be encoded as 6-bit ASCII. ipmitool
assumes these strings are only UTF/Latin1 encoded. Enabling 6-bit
encoding allows the sensor names to present up to 21 characaters,
instead of 16.

The OpenBMC project is introducing a change that allows the sensor
names to be extended to use the full 31 bytes of the Type/Len
encoding. OpenBMC is also enabling 6-bit ASCII encoding of the sensor
names along with the full 31 bytes of the string name. This allows
OpenBMC to express sensor names up to 41 characters in length.

The IPMI specification says SDR strings are 16 bytes maximum. It also
states any extra characters beyond 16 *should* be ignored. It does not
say they *must* be ignored. OpenBMC has been creating workarounds that
do not work well to live within the 16 character limit. Allowing 41
character string names is expected to be satisfactory.

Resolves: Allows sensor names to use 6-bit ASCII encoding
Signed-off-by: Johnathan Mantey <johnathanx.mantey@intel.com>