lan: Clean compile time warnings from LAN authentication
The MD2 and MD5 encryption schemes are deprecated. The compile time
switches, assigned from config.h, do not change which section of code
is compiled. The #ifdef is always seen as TRUE.

It is preferred to use #if to select the code to be compiled. Now
changing the HAVE_CRYPTO_MDx entries in the config.h file perform the
correct code compilation.

There are also two unused include files, which have been removed.
Resolves ipmitool/ipmitool/#21
Signed-off-by: Johnathan Mantey <johnathanx.mantey@intel.com>