Use _PATH_VARRUN instead of _PATH_RUN.
The _PATH_VARRUN is not specified in POSIX but is a de facto standard that
is defined on most systems.

There is no system that I know of that would have _PATH_RUN defined in
paths.h or anywhere else.  Checking #ifdef _PATH_RUN would  basically
hardcode into "/run" on any system.  Following FHS 3.0 is fine as long as
you respect build on other systems.

Fixes:	cdac4e07f7daeb6d05fa8018e8db76e96725998d